+++ src/components/page-header.tsx
@@ -173,7 +173,7 @@ export function PageHeader({
                 goBackButtonVariant === "ghost" &&
                   "hover:bg-gray-200 dark:hover:bg-gray-800 hover:ring-1 ring-inset ring-gray-300 dark:ring-gray-700",
                 goBackButtonVariant === "tinted" &&
-                  "backdrop-blur-xs hover:text-secondary-foreground/95 bg-gray-300/75 hover:bg-gray-300 dark:bg-gray-800/75 dark:hover:bg-gray-800 ring-1 ring-inset ring-gray-400/25 hover:ring-gray-400/80 dark:ring-gray-700/35 dark:hover:ring-gray-700",
+                  "backdrop-blur-sm hover:text-secondary-foreground/95 bg-gray-300/75 hover:bg-gray-300 dark:bg-gray-800/75 dark:hover:bg-gray-800 ring-1 ring-inset ring-gray-400/25 hover:ring-gray-400/80 dark:ring-gray-700/35 dark:hover:ring-gray-700",
                 "ml-3 shrink-0 size-7.5 transition-all duration-150 active-view-transition:transition-none",
                 isAtTop ? hoverAnimClasses.hidden : hoverAnimClasses.visible,
               )}
